backend/requirements-ai.txt
-r requirements.txt

# Ссылка на версию с CUDA с github
llama-cpp-python @ https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.4-cu121/llama_cpp_python-0.3.4-cp310-cp310-win_amd64.whl